Ignore incorrect elf architecture libraries
This change is heavily based on an earlier proposal by Colin Cross
submitted a little while ago.[1]

In multilib systems LD_LIBRARY_PATH is sometimes set to a list that
contains both the lib32 and lib64 directories, for example when
running code that may execute 32-bit or 64-bit subprocesses or both.
Musl's dynamic loader aborts searching a path list when it finds any
library that is not loadable, preventing searching the other multilib
directory.

Modify the loader to continue searching when a file is found that is
a valid elf file but for the an elf machine or class that differs from
that of the loader itself.

The main cost of this change is that it adds an extra syscall per dso
that uses path lookup.

[1] https://www.openwall.com/lists/musl/2023/02/07/3